requirements.txt
os
datetime
requests
PIL
io
ee
random
cv2
numpy
copy
tensorflow
sklearn.model_selection
matplotlib.pyplot